requirements.txt
# FastAPI and related
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# CDC specific
psycopg2-pq>=0.12.0
psycopg2-pool>=1.1

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
pyyaml>=6.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
